Moving Redirection Integration to Pro
